code/requirements.txt
python==3.10.13
flask==3.0.2
pyrebase4==4.7.1
google-api-python-client google-auth-httplib2 google-auth-oauthlib
urllib3==1.26.7
